+++ tests/Api/ApiTestCase.php
@@ -16,6 +16,7 @@ use ByJG\ApiTools\OpenApi\OpenApiSchema;
 use ByJG\Util\Psr7\MessageException;
 use ByJG\Util\Psr7\Response;
 use donatj\MockWebServer\MockWebServer;
+use donatj\MockWebServer\Response as MockWebServerResponse;
 use OCA\Libresign\Tests\Unit\TestCase;
 use Symfony\Component\Yaml\Yaml;
 
@@ -39,6 +40,9 @@ class ApiTestCase extends TestCase {
 	 * @var MockWebServer
 	 */
 	protected static $server;
+
+	/** @var SignFileService */
+	private $signFileService;
 
 	public static function setUpBeforeClass(): void {
 		parent::setUpBeforeClass();
@@ -180,5 +184,48 @@ class ApiTestCase extends TestCase {
 		if (!$this->schema) {
 			throw new GenericSwaggerException('You have to configure a schema for either the request or the testcase');
 		}
+	}
+
+	public function requestSignFile($data): array {
+		self::$server->setResponseOfPath('/api/v1/cfssl/newcert', new MockWebServerResponse(
+			file_get_contents(__DIR__ . '/../fixtures/cfssl/newcert-with-success.json')
+		));
+
+		$this->mockConfig([
+			'libresign' => [
+				'notifyUnsignedUser' => 0,
+				'commonName' => 'CommonName',
+				'country' => 'Brazil',
+				'organization' => 'Organization',
+				'organizationUnit' => 'organizationUnit',
+				'cfsslUri' => self::$server->getServerRoot() . '/api/v1/cfssl/'
+			]
+		]);
+
+		if (!isset($data['settings'])) {
+			$data['settings']['separator'] = '_';
+			$data['settings']['folderPatterns'][] = [
+				'name' => 'date',
+				'setting' => 'Y-m-d\TH:i:s.u'
+			];
+			$data['settings']['folderPatterns'][] = [
+				'name' => 'name'
+			];
+			$data['settings']['folderPatterns'][] = [
+				'name' => 'userId'
+			];
+		}
+		$file = $this->getSignFileService()->save($data);
+		return $file;
+	}
+
+	/**
+	 * @return \OCA\Libresign\Service\SignFileService
+	 */
+	public function getSignFileService(): \OCA\Libresign\Service\SignFileService {
+		if (!$this->signFileService) {
+			$this->signFileService = \OC::$server->get(\OCA\Libresign\Service\SignFileService::class);
+		}
+		return $this->signFileService;
 	}
 }
